Reorganization of mechanics dir.
