+++ src/morphac/mechanics/models/include/dubin_model.h
@@ -1,0 +1,9 @@
+#ifndef DUBIN_MODEL_H
+#define DUBIN_MODEL_H
+
+namespace morphac {
+namespace mechanics {
+
+} // namespace constructs
+} // namespace morphac
+
